Move App code into /src and use .jsx extension
- To match what is being done in besafe-app
- Code in /src will leave the root folder less cluttered, which is nice
- Technically, using HTML tags directly in the file, like there is in App.js
  and App-test.jsx is not stsandard javascript. This makes that difference
  more clear
