+++ index.js
@@ -1,9 +1,12 @@
-/**
- * @format
- */
+// Third-party dependencies
+import {
+    AppRegistry
+} from 'react-native'
 
-import {AppRegistry} from 'react-native';
-import App from './App';
-import {name as appName} from './app.json';
+// In-house dependencies
+import App from './src/App'
+import {
+    name as appName
+} from './app.json'
 
-AppRegistry.registerComponent(appName, () => App);
+AppRegistry.registerComponent(appName, () => App)
